requirements.txt
streamlit>=1.39
ultralytics>=8.0
plotly>=5.20
scikit-learn>=1.3
scipy>=1.10
opencv-python>=4.8
pillow>=10.0
numpy>=1.24
pandas>=2.0

# Development and testing dependencies
pytest>=7.0
pytest-cov>=4.0
black>=23.0
flake8>=6.0
mypy>=1.0
